requirements.txt
astroid==2.2.4
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coverage==4.5.2
coveralls==1.6.0
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
isort==4.3.10
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
